requirements.txt
# medicaid mco rfp analysis pipeline
# python 3.11+

# data processing
pandas>=2.2.0
numpy>=1.26.0

# document extraction
pypdf>=5.9.0
python-docx>=1.2.0

# analysis
scipy>=1.12.0
scikit-learn>=1.4.0

# llm integration (optional - for full pipeline)
anthropic>=0.72.0

# visualization (optional)
matplotlib>=3.8.0
seaborn>=0.13.0

# progress tracking
tqdm>=4.66.0

# testing
pytest>=8.0.0
